+++ activesupport/lib/active_support/cache/file_store.rb
@@ -13,8 +13,7 @@ module ActiveSupport
       attr_reader :cache_path
 
       DIR_FORMATTER = "%03X"
-      FILENAME_MAX_SIZE = 228 # max filename size on file system is 255, minus room for timestamp and random characters appended by Tempfile (used by atomic write)
-      FILEPATH_MAX_SIZE = 900 # max is 1024, plus some room
+      FILENAME_MAX_SIZE = 113 # max filename size on file system is 140, minus room for timestamp and random characters appended by Tempfile (used by atomic write) http://unix.stackexchange.com/questions/32795/what-is-the-maximum-allowed-filename-and-folder-size-with-ecryptfs/32834#32834
       EXCLUDED_DIRS = ['.', '..'].freeze
 
       def initialize(cache_path, options = nil)
@@ -118,7 +117,7 @@ module ActiveSupport
 
         # Translate a key into a file path.
         def key_file_path(key)
-          if key.size > FILEPATH_MAX_SIZE
+          if key.size > FILENAME_MAX_SIZE
             key = Digest::MD5.hexdigest(key)
           end
 
@@ -126,15 +125,8 @@ module ActiveSupport
           hash = Zlib.adler32(fname)
           hash, dir_1 = hash.divmod(0x1000)
           dir_2 = hash.modulo(0x1000)
-          fname_paths = []
 
-          # Make sure file name doesn't exceed file system limits.
-          begin
-            fname_paths << fname[0, FILENAME_MAX_SIZE]
-            fname = fname[FILENAME_MAX_SIZE..-1]
-          end until fname.blank?
-
-          File.join(cache_path, DIR_FORMATTER % dir_1, DIR_FORMATTER % dir_2, *fname_paths)
+          File.join(cache_path, DIR_FORMATTER % dir_1, DIR_FORMATTER % dir_2, fname)
         end
 
         # Translate a file path into a key.